# specs/001-iot-gas-detection/contracts/api.yaml
openapi: 3.0.3
info:
  title: Sistema IoT de Detección de Fugas de Gas API
  description: |
    API REST para monitoreo y control del sistema de detección de fugas de gas.
    Implementa RBAC con 4 roles: admin, operator, technician, auditor.
  version: 1.0.0
  contact:
    name: GEUS Development Team
servers:
  - url: https://{domain}/api/v1
    description: Production server
    variables:
      domain:
        default: api.geus.example.com
security:
  - bearerAuth: []
...
paths:
  /auth/login:
    post:
      tags: [Auth]
      summary: User login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Successful login
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: Token refreshed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /alerts:
    get:
      tags: [Alerts]
      summary: List alerts
      parameters:
        - name: sensor_id
          in: query
          schema:
            type: string
            format: uuid
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/AlertStatus'
        - name: severity
          in: query
          schema:
            $ref: '#/components/schemas/AlertSeverity'
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: List of alerts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlertListResponse'
  /alerts/{alert_id}:
    parameters:
      - name: alert_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      tags: [Alerts]
      summary: Get alert details
      responses:
        '200':
          description: Alert details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Alert'
  /alerts/{alert_id}/acknowledge:
    parameters:
      - name: alert_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      tags: [Alerts]
      summary: Acknowledge alert
      description: Admin, Operator (assigned sensors only)
      responses:
        '200':
          description: Alert acknowledged
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Alert'
  /alerts/{alert_id}/resolve:
    parameters:
      - name: alert_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      tags: [Alerts]
      summary: Resolve alert
      description: |
        Admin, Operator (assigned sensors only).
        Critical alerts require explicit resolution.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                notes:
                  type: string
                  maxLength: 500
      responses:
        '200':
          description: Alert resolved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Alert'
...
  /commands/dissipator/{sensor_id}:
    parameters:
      - $ref: '#/components/parameters/SensorIdParam'
    post:
      tags: [Commands]
      summary: Control dissipator
      description: |
        Admin, Operator (assigned sensors).
        Cannot turn off during active gas alert.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DissipatorCommand'
      responses:
        '202':
          description: Command accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommandResponse'
        '409':
          description: Cannot turn off during active alert
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /commands/test-mode/{sensor_id}:
    parameters:
      - $ref: '#/components/parameters/SensorIdParam'
    post:
      tags: [Commands]
      summary: Toggle test mode
      description: Technician only. Auto-expires after 30 minutes.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TestModeCommand'
      responses:
        '202':
          description: Command accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommandResponse'
...
  /audit/logs:
    get:
      tags: [Audit]
      summary: Query audit logs
      description: Auditor only (read-only)
      parameters:
        - name: start
          in: query
          schema:
            type: string
            format: date-time
        - name: end
          in: query
          schema:
            type: string
            format: date-time
        - name: user_id
          in: query
          schema:
            type: string
            format: uuid
        - name: action
          in: query
          schema:
            type: string
        - name: sensor_id
          in: query
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: Audit logs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditLogListResponse'
  /audit/export:
    post:
      tags: [Audit]
      summary: Export audit data
      description: Auditor only. Async export for forensic analysis.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExportRequest'
      responses:
        '202':
          description: Export started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExportResponse'
  /audit/security:
    get:
      tags: [Audit]
      summary: Security metrics
      description: Auditor only. Failed logins, anomalies, latency.
      parameters:
        - name: period
          in: query
          required: true
          schema:
            type: string
            enum: [1h, 6h, 24h, 7d]
      responses:
        '200':
          description: Security metrics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SecurityMetrics'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    SensorIdParam:
      name: sensor_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  responses:
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
    RefreshRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token:
          type: string
    TokenResponse:
      type: object
      required: [access_token, refresh_token, expires_in]
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        expires_in:
          type: integer
          description: Seconds until expiration
        user:
          $ref: '#/components/schemas/User'
...
    AlertStatus:
      type: string
      enum: [active, acknowledged, resolved]
    AlertSeverity:
      type: string
      enum: [warning, critical]
    UserRole:
      type: string
      enum: [admin, operator, technician, auditor]
...
    Alert:
      type: object
      properties:
        id:
          type: string
          format: uuid
        sensor_id:
          type: string
          format: uuid
        sensor_location:
          type: string
        gas_level_ppm:
          type: number
        severity:
          $ref: '#/components/schemas/AlertSeverity'
        status:
          $ref: '#/components/schemas/AlertStatus'
        triggered_at:
          type: string
          format: date-time
        acknowledged_at:
          type: string
          format: date-time
          nullable: true
        acknowledged_by:
          type: string
          format: uuid
          nullable: true
        resolved_at:
          type: string
          format: date-time
          nullable: true
        resolved_by:
          type: string
          format: uuid
          nullable: true
        auto_resolved:
          type: boolean
    AlertListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Alert'
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
...
    DissipatorCommand:
      type: object
      required: [action]
      properties:
        action:
          type: string
          enum: [on, off]
    TestModeCommand:
      type: object
      required: [enabled]
      properties:
        enabled:
          type: boolean
        duration_minutes:
          type: integer
          minimum: 5
          maximum: 60
          default: 30
    CommandResponse:
      type: object
      properties:
        command_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [accepted, pending, delivered, failed]
        message:
          type: string
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        name:
          type: string
        role:
          $ref: '#/components/schemas/UserRole'
        status:
          type: string
          enum: [active, suspended]
        notifications_enabled:
          type: boolean
        last_login_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
...
    AuditLog:
      type: object
      properties:
        id:
          type: string
          format: uuid
        timestamp:
          type: string
          format: date-time
        user_id:
          type: string
          format: uuid
          nullable: true
        user_email:
          type: string
          nullable: true
        user_role:
          $ref: '#/components/schemas/UserRole'
        action:
          type: string
        sensor_id:
          type: string
          format: uuid
          nullable: true
        details:
          type: object
        ip_origin:
          type: string
    AuditLogListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/AuditLog'
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    ExportRequest:
      type: object
      required: [type, start, end]
      properties:
        type:
          type: string
          enum: [audit_logs, telemetry, alerts]
        start:
          type: string
          format: date-time
        end:
          type: string
          format: date-time
        sensor_ids:
          type: array
          items:
            type: string
            format: uuid
        format:
          type: string
          enum: [csv, json]
          default: csv
    ExportResponse:
      type: object
      properties:
        export_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [pending, processing, completed, failed]
        download_url:
          type: string
          nullable: true
        expires_at:
          type: string
          format: date-time
          nullable: true
    SecurityMetrics:
      type: object
      properties:
        period:
          type: string
        failed_logins:
          type: integer
        failed_mqtt_connections:
          type: integer
        anomalous_traffic_events:
          type: integer
        avg_api_latency_ms:
          type: number
        avg_mqtt_latency_ms:
          type: number
...